pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kr-enforcement-cases"
version = "1.0.0"
description = "Korean FSS/SFC enforcement case dataset — scraping, enrichment, Beneish validation, and structured violation taxonomy"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "pdfplumber>=0.10",
    "pypdfium2>=4.0",
    "anthropic>=0.30",
    "pydantic>=2",
    "python-dotenv>=1",
    "opendartreader>=0.2",
    "rapidfuzz>=3.0",
    "python-hwpx>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.uv]
compile-bytecode = false

[tool.pytest.ini_options]
testpaths = ["tests"]
